feat(frontend): scheduled runs extracted to @agenta/settings-ui
Second of the three Triggers sections. Same conversion as the connections list:
DataTable, @agenta/ui primitives, no antd.

The cron column's hover tooltip becomes a title attribute — a Radix tooltip per
row is a popper per row, and the raw cron string is a hint, not a control.

Triggers.tsx imports this one from the package directly rather than through a
re-export stub at the old path, which OSS lint forbids.

Gates: lint 24/24, tsc 0 across settings-ui, oss, ee and mobile.
