+++ web/oss/src/components/pages/settings/Triggers/Triggers.tsx
@@ -1,6 +1,6 @@
 import {TriggerDeliveriesDrawer} from "@agenta/entity-ui/gatewayTrigger"
+import {TriggerSchedulesSection} from "@agenta/settings-ui"
 
-import GatewaySchedulesSection from "./components/GatewaySchedulesSection"
 import GatewaySubscriptionsSection from "./components/GatewaySubscriptionsSection"
 import GatewayTriggersSection from "./components/GatewayTriggersSection"
 
@@ -10,7 +10,7 @@ export default function Triggers() {
         <div className="flex flex-col">
             <GatewayTriggersSection />
             <GatewaySubscriptionsSection />
-            <GatewaySchedulesSection />
+            <TriggerSchedulesSection />
             {/* One shared deliveries drawer for both subscriptions and schedules
                 (both bind the same atom; rendering it once avoids a duplicate). */}
             <TriggerDeliveriesDrawer />
